CLN: Moved members listing to list partials
